Add create Slack notificator modal box in Gamecraft UI

+++ gamecraft-ui/src/main/resources/static/js/ui_operations.js
@@ -221,15 +221,28 @@ function fillSlackAccountTable() {
     var bots = getSlackAccounts();
     $(function() {
         $.each(bots, function(i, item) {
-            var tr = $('<tr>').append(
-                $('<tr>'),
-                $('<td>').text(item.id),
-                $('<td>').text(item.slackAccountName),
-                $('<td>').text(item.slackAccountDescription),
-                $('<td>').html("<button type=\"button\" class=\"btn btn-danger btn-xs\" onclick=\"deleteSlackAccount('" +item.id + "'); location.reload(); \"><span class=\"glyphicon glyphicon-remove\"></span> </button>")
-            );
-            $(".table").append(tr.html());
+            if (isAdmin(getUsername())) {
+                var tr = $('<tr>').append(
+                    $('<tr>'),
+                    $('<td>').text(item.id),
+                    $('<td>').text(item.slackAccountName),
+                    $('<td>').text(item.slackAccountDescription),
+                    $('<td>').html("<button type=\"button\" class=\"btn btn-danger btn-xs\" onclick=\"deleteSlackAccount('" + item.id + "'); location.reload(); \"><span class=\"glyphicon glyphicon-remove\"></span> </button>")
+                );
+                $(".table").append(tr.html());
+            }
+            else {
+                var tr = $('<tr>').append(
+                    $('<tr>'),
+                    $('<td>').text(item.id),
+                    $('<td>').text(item.slackAccountName),
+                    $('<td>').text(item.slackAccountDescription),
+                    $('<td>').html("")
+                );
+                $(".table").append(tr.html());
+            }
         });
+
     });
 }
